Move area_to_polygon to subpackage geometry
